+++ src/main/groovy/com/getkeepsafe/dexcount/DexMethodCountTask.groovy
@@ -31,6 +31,12 @@ import org.gradle.logging.StyledTextOutput
 import org.gradle.logging.StyledTextOutputFactory
 
 class DexMethodCountTask extends DefaultTask {
+    /**
+     * The maximum number of method refs and field refs allowed in a single Dex
+     * file.
+     */
+    private static final int MAX_DEX_REFS = 0xFFFF;
+
     def PackageTree tree;
 
     @Input
@@ -64,6 +70,8 @@ class DexMethodCountTask extends DefaultTask {
         withStyledOutput(StyledTextOutput.Style.Info) { out ->
             out.println("Total methods in ${filename}: ${tree.methodCount}")
             out.println("Total fields in ${filename}:  ${tree.fieldCount}")
+            out.println("Methods remaining in ${filename}: ${MAX_DEX_REFS - tree.methodCount}")
+            out.println("Fields remaining in ${filename}:  ${MAX_DEX_REFS - tree.fieldCount}")
         }
 
         if (summaryFile != null) {